requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
cffi==1.17.1
chargily_pay==2.1.3
charset-normalizer==3.4.0
cloudflare==3.1.1
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.4
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
model-bakery==1.20.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.22.3
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
stripe==11.3.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
